requirements.txt
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.1
dateparser==0.7.6
Flask==1.1.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.1
plotly==4.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-binance==0.7.5
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
retrying==1.3.3
service-identity==18.1.0
six==1.15.0
ta==0.5.25
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
zope.interface==5.1.0
